+++ src/index.ts
@@ -2,9 +2,11 @@ export * from './FileTypeIcon';
 export * from './ListView';
 export * from './Placeholder';
 export * from './SiteBreadcrumb';
+export * from './TaxonomyPicker';
 export * from './WebPartTitle';
-export * from './IFrameDialog';
+export * from './ListPicker';
 
+export * from './IFrameDialog';
 export * from './Common';
 export * from './Utilities';
 export * from './IFrameDialog';